fix(cellguide): server side props for seo description (#5832)
## Reason for Change

- CellGuide pipeline was updated to generate SEO descriptions as
strings, not dictionaries. Because SEO descriptions are only generated
when new cell types are added, most of the SEO descriptions are still
dictionaries. Newer cell types (like [this
one](https://cellxgene.cziscience.com/cellguide/CL_4030026)) store the
SEO descriptions as strings, leading to an error in parsing the server
side props and undefined names.
- This PR updates the frontend server to get the cell type names from
the `celltype_metadata.json` and updates the CellGuide query module to
expect strings for the GPT SEO description.
- Once this is deployed, the old GPT SEO descriptions will be migrated
to the correct format (strings).